Sizes of cursors and pages are integers in REST.
This change addresses the problem of uint64 not being representable in
javascript, leading to a swagger definition of two numbers related to
cursors, namely `size` and `total_records`, having string type rather
than integer. Using a uint32 generates the desired type and is
sufficient for paging purposes.

This is a breaking change, which is why this commit contains changes
to CLI as well. For the record, the feature is currently under feature
flag.

Fixes #3934

+++ cmd/cli/app/history/history_list.go
@@ -66,7 +66,7 @@ func listCommand(ctx context.Context, cmd *cobra.Command, _ []string, conn *grpc
 
 	// page options
 	cursorStr := viper.GetString("cursor")
-	size := viper.GetUint64("size")
+	size := viper.GetUint32("size")
 
 	format := viper.GetString("output")
 
@@ -141,7 +141,7 @@ func listCommand(ctx context.Context, cmd *cobra.Command, _ []string, conn *grpc
 	return nil
 }
 
-func cursorFromOptions(cursorStr string, size uint64) *minderv1.Cursor {
+func cursorFromOptions(cursorStr string, size uint32) *minderv1.Cursor {
 	var cursor *minderv1.Cursor
 	if cursorStr != "" || size != 0 {
 		cursor = &minderv1.Cursor{}
